builder/requirements.txt
azure-cosmos
pandas
pyarrow
numpy
aiohttp